Move validators into machine module
