+++ gordo_components/machine/validators.py
