+++ gordo_components/workflow/config_elements/normalized_config.py
@@ -2,7 +2,7 @@
 
 from typing import List
 
-from gordo_components.workflow.config_elements.validators import fix_runtime
+from gordo_components.machine.validators import fix_runtime
 from gordo_components.workflow.workflow_generator.helpers import patch_dict
 from gordo_components.machine import Machine
